Move pure editor utilities from components/ to lib/editor/
Move instructionBlockUtils, preprocessMarkdown and ListItemExtension
to front/lib/editor/ and update all import paths.
