+++ front/components/poke/pages/AssistantInstructionsPage.tsx
@@ -1,6 +1,6 @@
 import { buildAgentInstructionsReadOnlyExtensions } from "@app/components/agent_builder/instructions/AgentBuilderInstructionsEditor";
-import { preprocessMarkdownForEditor } from "@app/components/editor/lib/preprocessMarkdownForEditor";
 import { useWorkspace } from "@app/lib/auth/AuthContext";
+import { preprocessMarkdownForEditor } from "@app/lib/editor/preprocessMarkdown";
 import { useRequiredPathParam } from "@app/lib/platform";
 import { usePokeAgentDetails } from "@app/poke/swr/agent_details";
 import { Spinner } from "@dust-tt/sparkle";
