+++ front/lib/editor/instructionBlockUtils.ts
